// src/vs/platform/agentHost/node/codex/protocol/generated/ClientNotification.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`
// Generated from @openai/codex 0.142.0, licensed Apache-2.0.
// Modified by VS Code to rewrite imports and apply repository formatting.
// See README.md in this directory for provenance and licensing notes.
// Regenerate: npm run codex:gen-protocol
// Source file: ClientNotification.ts

export type ClientNotification = { "method": "initialized" };